kiwi: Update kernel & dt.img from a traceable source
The last kernel was not from an official distribution and source
was not provided.  It contained a version of f2fs that included
(incorrect) manual patching.

This kernel is built from:

https://github.com/LineageOS/android_kernel_huawei_kiwi/<commit>

TODO: Fix the link when are done testing the changes and the f2fs
support is merged to lineage.

The f2fs implementation matches the commit 497842998b07f411103f96759680f971285eb7c3
of f2fs-stable.

Change-Id: Ia8b45843948c3ae95192a367798164cd418a7b7e