kiwi: Remove build instructions
They are not correct (do not point to the correct repo for
libcryptfw_hw, for example) and it's probably not a good idea
to try to reproduce build instructions in each device repo.

If twrp wants to have build instructions then they should be
updated on the website, not in a device tree.

Change-Id: I14a5a0401f207db2a5913fe83ec3925f7c38db12